test: Create macro for testing purpose
