+++ interpreter/ezcript_ast/src/lib.rs
@@ -2,18 +2,8 @@ mod ast;
 mod expressions;
 mod statements;
 
-// TODO: Convert this into a macro creating a lib just for macros and also create a macro for know
-// if a value is not None
-// fn is_instance<U, T>(_: &U, _: T) -> bool {
-//     let value_type = std::any::type_name::<U>();
-//     let the_type = std::any::type_name::<T>();
-
-//     if value_type != the_type {
-//         panic!("The value type of the two arguments are not the same!");
-//     }
-
-//     true
-// }
+#[macro_use]
+extern crate utils;
 
 #[cfg(test)]
 mod test {
@@ -32,29 +22,4 @@ mod test {
     //         panic!("This value shouldn't be None");
     //     }
     // }
-
-    #[derive(Debug)]
-    struct Codo {
-        codo: String,
-        num: u64,
-    }
-
-    impl Codo {
-        pub fn new() -> Self {
-            Self {
-                codo: "".to_string(),
-                num: 0,
-            }
-        }
-    }
-
-    #[test]
-    fn testing() {
-        let codo = Codo {
-            codo: "mi codo".to_string(),
-            num: 42,
-        };
-
-        is_instance(&codo, Codo::new());
-    }
 }
